Fixed Gpio and added a working OLED driver sample
OLED is based on SSD1306 and is a UG-2864HSWEG01 device, 128 x 64. The
driver holds a live copy of the ram and has a 5 x 7 character font which
you can just pass strings to. There's still more to do with graphics and
single reliable screen writes, I think noise is my issue though!